Moved LeftButton and RightButton into components and modified their imports
